refactor: do minor notification settings dialog refactoring
