+++ src/features/notifications/components/notification-setting-modal/types.ts
@@ -1,0 +1,11 @@
+export interface NotificationSetting {
+  email: boolean;
+  inApp: boolean;
+  listingScope: 'mine' | 'all';
+}
+
+export enum AlertCategory {
+  SPONSOR = 'SPONSOR',
+  TALENT = 'TALENT',
+  GENERAL = 'GENERAL',
+}
